// client_app/src/lib/generated/osmosis/osmosis/poolmanager/v1beta1/swap_route.ts
export interface SwapAmountInRoute {
  poolId: number;
  tokenOutDenom: string;
}

export interface SwapAmountOutRoute {
  poolId: number;
  tokenInDenom: string;
}

export interface SwapAmountInSplitRoute {
  pools: SwapAmountInRoute[];
  tokenInAmount: string;
}

export interface SwapAmountOutSplitRoute {
  pools: SwapAmountOutRoute[];
  tokenOutAmount: string;
}

function createBaseSwapAmountInRoute(): SwapAmountInRoute {
  return { poolId: 0, tokenOutDenom: "" };
}

export const SwapAmountInRoute: MessageFns<SwapAmountInRoute> = {
  encode(message: SwapAmountInRoute, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.poolId !== 0) {
      writer.uint32(8).uint64(message.poolId);
    }
    if (message.tokenOutDenom !== "") {
      writer.uint32(18).string(message.tokenOutDenom);
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): SwapAmountInRoute {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseSwapAmountInRoute();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 8) {
            break;
          }

          message.poolId = longToNumber(reader.uint64());
          continue;
        }
        case 2: {
          if (tag !== 18) {
            break;
          }

          message.tokenOutDenom = reader.string();
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): SwapAmountInRoute {
    return {
      poolId: isSet(object.poolId)
        ? globalThis.Number(object.poolId)
        : isSet(object.pool_id)
        ? globalThis.Number(object.pool_id)
        : 0,
      tokenOutDenom: isSet(object.tokenOutDenom)
        ? globalThis.String(object.tokenOutDenom)
        : isSet(object.token_out_denom)
        ? globalThis.String(object.token_out_denom)
        : "",
    };
  },

  toJSON(message: SwapAmountInRoute): unknown {
    const obj: any = {};
    if (message.poolId !== 0) {
      obj.poolId = Math.round(message.poolId);
    }
    if (message.tokenOutDenom !== "") {
      obj.tokenOutDenom = message.tokenOutDenom;
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<SwapAmountInRoute>, I>>(base?: I): SwapAmountInRoute {
    return SwapAmountInRoute.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<SwapAmountInRoute>, I>>(object: I): SwapAmountInRoute {
    const message = createBaseSwapAmountInRoute();
    message.poolId = object.poolId ?? 0;
    message.tokenOutDenom = object.tokenOutDenom ?? "";
    return message;
  },
};
// ...
export const SwapAmountInSplitRoute: MessageFns<SwapAmountInSplitRoute> = {
  encode(message: SwapAmountInSplitRoute, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    for (const v of message.pools) {
      SwapAmountInRoute.encode(v!, writer.uint32(10).fork()).join();
    }
    if (message.tokenInAmount !== "") {
      writer.uint32(18).string(message.tokenInAmount);
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): SwapAmountInSplitRoute {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseSwapAmountInSplitRoute();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 10) {
            break;
          }

          message.pools.push(SwapAmountInRoute.decode(reader, reader.uint32()));
          continue;
        }
        case 2: {
          if (tag !== 18) {
            break;
          }

          message.tokenInAmount = reader.string();
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): SwapAmountInSplitRoute {
    return {
      pools: globalThis.Array.isArray(object?.pools) ? object.pools.map((e: any) => SwapAmountInRoute.fromJSON(e)) : [],
      tokenInAmount: isSet(object.tokenInAmount)
        ? globalThis.String(object.tokenInAmount)
        : isSet(object.token_in_amount)
        ? globalThis.String(object.token_in_amount)
        : "",
    };
  },

  toJSON(message: SwapAmountInSplitRoute): unknown {
    const obj: any = {};
    if (message.pools?.length) {
      obj.pools = message.pools.map((e) => SwapAmountInRoute.toJSON(e));
    }
    if (message.tokenInAmount !== "") {
      obj.tokenInAmount = message.tokenInAmount;
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<SwapAmountInSplitRoute>, I>>(base?: I): SwapAmountInSplitRoute {
    return SwapAmountInSplitRoute.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<SwapAmountInSplitRoute>, I>>(object: I): SwapAmountInSplitRoute {
    const message = createBaseSwapAmountInSplitRoute();
    message.pools = object.pools?.map((e) => SwapAmountInRoute.fromPartial(e)) || [];
    message.tokenInAmount = object.tokenInAmount ?? "";
    return message;
  },
};
// ...
type Builtin = Date | Function | Uint8Array | string | number | boolean | undefined;

export type DeepPartial<T> = T extends Builtin ? T
  : T extends globalThis.Array<infer U> ? globalThis.Array<DeepPartial<U>>
  : T extends ReadonlyArray<infer U> ? ReadonlyArray<DeepPartial<U>>
  : T extends {} ? { [K in keyof T]?: DeepPartial<T[K]> }
  : Partial<T>;

type KeysOfUnion<T> = T extends T ? keyof T : never;
export type Exact<P, I extends P> = P extends Builtin ? P
  : P & { [K in keyof P]: Exact<P[K], I[K]> } & { [K in Exclude<keyof I, KeysOfUnion<P>>]: never };
// ...
export interface MessageFns<T> {
  encode(message: T, writer?: BinaryWriter): BinaryWriter;
  decode(input: BinaryReader | Uint8Array, length?: number): T;
  fromJSON(object: any): T;
  toJSON(message: T): unknown;
  create<I extends Exact<DeepPartial<T>, I>>(base?: I): T;
  fromPartial<I extends Exact<DeepPartial<T>, I>>(object: I): T;
}